dynamicNewsletter/requirements.txt
streamlit
json
calendar
datetime
google.auth.transport.requests
google.oauth2.credentials
google_auth_oauthlib.flow
googleapiclient.discovery
googleapiclient.errors
